wip: split campaign distribution source resolution
